backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pydantic-settings
httpx[http2]
python-jose[cryptography]
apscheduler
aiosmtplib
redis
email-validator 
langchain
langchain-core
grpcio
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0
tiktoken>=0.7.0